refactor. update home page comp
